Issue: Haciendo que enemigo sea Interfaz. Adaptacion del codigo.
