+++ src/main/java/edu/fiuba/algo3/modelo/enemigos/Topo.java
@@ -1,23 +1,45 @@
 package edu.fiuba.algo3.modelo.enemigos;
 
-import edu.fiuba.algo3.modelo.estados.EstadoVivo;
 import edu.fiuba.algo3.modelo.juego.Jugador;
+import edu.fiuba.algo3.modelo.mapa.Mapa;
+import edu.fiuba.algo3.modelo.parcelas.Parcela;
 import edu.fiuba.algo3.modelo.parcelas.ParcelaDePasarela;
+import edu.fiuba.algo3.modelo.parcelas.TerrenoNoAptoParaCaminar;
+import edu.fiuba.algo3.modelo.parcelas.TerrenoNoAptoParaConstruir;
 
-public class Topo extends Enemigo {
-
+public class Topo implements Enemigo {
     private Atacante atacante;
+    private Trasladable trasladable;
 
     public Topo(ParcelaDePasarela pasarela) {
-        this.creditos = 1;
-       // this.danio = 2;
-        this.estadoDeVida = new EstadoVivo(5);
-        this.pasarelaActual = pasarela;
         this.trasladable = new BajoTierra(1, pasarela, 0);
         this.atacante = new AtacanteDeJugador(2);
     }
 
     public void atacar(Jugador jugador) {
         this.atacante.atacar(jugador);
+    }
+
+    public boolean puedeMoverseA(Parcela parcela) {
+        return (parcela.puedeMoverseAqui());
+    }
+    public ParcelaDePasarela getPasarelaActual() {
+        return (trasladable.getPasarelaActual());
+    }
+
+    public void setPasarelaActual(ParcelaDePasarela pasarela) {
+        this.trasladable.setPasarelaActual(pasarela);
+    }
+
+    public void mover(Mapa mapa) throws TerrenoNoAptoParaConstruir, TerrenoNoAptoParaCaminar {
+        this.trasladable = trasladable.moverse(mapa);
+
+    }
+    @Override
+    public void recibirDanio(int puntosARecibir) {
+
+    }
+    @Override
+    public void recibirAtaque(Parcela parcelaDefensa, int rangoAtaque, int danio) throws EnemigoFueraDeRango {
     }
 }
